bindings: containers/containers uses entities/types
